+++ src/lib/st-search/st-search.component.ts
@@ -18,19 +18,43 @@ import {
    OnInit,
    Output,
    SimpleChanges,
-   Renderer,
+   Renderer2,
    ChangeDetectorRef,
    ViewChild,
    ElementRef
 } from '@angular/core';
 import { FormControl } from '@angular/forms';
 import { Subscription } from 'rxjs/Subscription';
-
 import 'rxjs/add/operator/debounceTime';
-
 import { StDropDownMenuItem } from '../st-dropdown-menu/st-dropdown-menu.interface';
 import { EventWindowManager } from '../utils/event-window-manager';
 
+/**
+ * @description {Component} [Search]
+ *
+ * The search component has been designed to allow user to find a specific content according to his needs.
+ *
+ * @example
+ *
+ * {html}
+ *
+ * ```
+ * Search without filters
+ *  <st-search [placeholder]="placeholder" [qaTag]="qaTag"
+ *  [value]="searched" [debounce]="debounceTime" [minLength]="minLength"
+ *  (search)="onSearchResult($event)"></st-search>
+ *
+ *
+ * Search with filters
+ *  <st-search [placeholder]="placeholder" [qaTag]="qaTag"
+ *  [value]="searched" [debounce]="debounceTime" [minLength]="minLength" [filterOptions]="[
+ *  { label: 'All', value: 1 },
+ *  { label: 'Type ', value: 2 },
+ *  { label: 'Color', value: 3 }
+ *  ]" (search)="onSearchResult($event)"></st-search>
+ * ```
+ *
+ */
 @Component({
    selector: 'st-search',
    templateUrl: './st-search.component.html',
@@ -38,31 +62,59 @@ import { EventWindowManager } from '../utils/event-window-manager';
    changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class StSearchComponent extends EventWindowManager implements OnChanges, OnDestroy, OnInit {
+   /** @Input {number} [debounce=0] Time elapsed in milliseconds before displaying the autocomplete list */
    @Input() debounce: number = 0;
+
+   /** @Input {boolean} [liveSearch=true] Boolean to enable or disable the automatic search while the user is typing */
    @Input() liveSearch: boolean = true;
+
+   /** @Input {number} [minLength=0] Minimum of characters typed by the user before launching the search */
    @Input() minLength: number = 0;
+
+   /** @Input {string} [placeholder='Search'] Text displayed in the search input */
    @Input() placeholder: string = 'Search';
+
+   /** @Input {string} [qaTag=''] Label used as id */
    @Input() qaTag: string;
+
+   /** @Input {string} [value=''] Initial value of the search text */
    @Input() value: string;
+
+   /** @Input {boolean} [disabled=false] Boolean to enable or disable the search */
    @Input() disabled: boolean = false;
 
+   /** @Input {boolean} [withAutocomplete=false] Enable or disable the autocomplete list when user is typing */
    @Input() withAutocomplete: boolean = false;
+
+   /** @Input {StDropDownMenuItem[]} [autocompleteList=''] List of items displayed in the autocomplete list when user is typing */
    @Input() autocompleteList: StDropDownMenuItem[] = [];
+
+   /** @Input {string} [emptyAutocompleteListMessage=''] Message displayed when the autocomplete list is enabled but
+    * there are not any item with the typed text
+    */
    @Input() emptyAutocompleteListMessage: string = '';
 
-   @Output() search: EventEmitter<string> = new EventEmitter<string>();
+   /** @Input {StDropDownMenuItem[]} [filterOptions=''] Options displayed at the filter select. If it is not introduced,
+    * filter will not be displayed
+    */
+   @Input() filterOptions: StDropDownMenuItem[];
+
+   /** @Output { Object(filter?: string, text: string)} [search=''] Event emitted when search is launched. It contains
+    * the text typed by the user and the filter value selected (only if filter is displayed)
+    */
+   @Output() search: EventEmitter<{filter?: string, text: string}> = new EventEmitter<{filter?: string, text: string}>();
 
    public searchBox: FormControl = new FormControl();
    public showClear: boolean;
+   public filter: string;
 
    private subscriptionSearch: Subscription | undefined = undefined;
    private subscriptionSearchClearButton: Subscription | undefined = undefined;
-   private lastEmited: string | undefined = undefined;
+   private lastEmittedText: string | undefined = undefined;
 
-   constructor(
-      private _render: Renderer,
-      private cd: ChangeDetectorRef,
-      @ViewChild('buttonId') public buttonElement: ElementRef) {
+   constructor(private _render: Renderer2,
+               private cd: ChangeDetectorRef,
+               @ViewChild('buttonId') public buttonElement: ElementRef) {
       super(_render, cd, buttonElement);
    }
 
@@ -78,13 +130,20 @@ export class StSearchComponent extends EventWindowManager implements OnChanges, 
       this.subscriptionSearchClearButton = this.searchBox.valueChanges.subscribe((val) => this.showClear = (val && val.length > 0));
       this.checkDisabled();
       this.manageSubscription();
+      if (this.filterOptions) {
+         this.filter = this.filterOptions[0].value;
+      }
    }
 
    public ngOnChanges(changes: SimpleChanges): void {
       this.checkDebounceChange(changes);
       this.checkValueChange(changes);
       this.checkDisableChange(changes);
-      this.checkAutocompleteMenuChange(changes);
+      this.checkAutoCompleteMenuChange(changes);
+   }
+
+   public onChangeFilter(): void {
+      this.emitValue(false);
    }
 
    public ngOnDestroy(): void {
@@ -98,8 +157,8 @@ export class StSearchComponent extends EventWindowManager implements OnChanges, 
    }
 
    public launchSearch(force: boolean): void {
-      if (this.canSearch(force)) {
-         this.showAutocompleteMenu();
+      if (this.canSearch()) {
+         this.showAutoCompleteMenu();
          this.emitValue(force);
       } else {
          this.closeElement();
@@ -132,12 +191,16 @@ export class StSearchComponent extends EventWindowManager implements OnChanges, 
 
    private emitValue(force: boolean): void {
       if (this.isEqualPrevious(force)) {
-         this.lastEmited = this.searchBox.value;
-         this.search.emit(this.lastEmited);
+         this.lastEmittedText = this.searchBox.value;
+         let newSearch: {filter?: string, text: string} = {text: this.lastEmittedText || ''};
+         if (this.filter) {
+            newSearch.filter = this.filter;
+         }
+         this.search.emit(newSearch);
       }
    }
 
-   private showAutocompleteMenu(): void {
+   private showAutoCompleteMenu(): void {
       if (this.withAutocomplete && !this.isActive) {
          this.openElement();
       }
@@ -155,7 +218,7 @@ export class StSearchComponent extends EventWindowManager implements OnChanges, 
       }
    }
 
-   private canSearch(force: boolean): boolean {
+   private canSearch(): boolean {
       return this.isDefined() && !this.disabled && this.checkMins();
    }
 
@@ -169,7 +232,7 @@ export class StSearchComponent extends EventWindowManager implements OnChanges, 
    }
 
    private isEqualPrevious(force: boolean): boolean {
-      return this.lastEmited !== this.searchBox.value || force;
+      return this.lastEmittedText !== this.searchBox.value || force;
    }
 
    private checkValueChange(changes: SimpleChanges): void {
@@ -194,7 +257,7 @@ export class StSearchComponent extends EventWindowManager implements OnChanges, 
       }
    }
 
-   private checkAutocompleteMenuChange(changes: SimpleChanges): void {
+   private checkAutoCompleteMenuChange(changes: SimpleChanges): void {
       if (changes && changes.autocompleteList) {
          this.cd.markForCheck();
       }